+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+## 6.1.0 - 2024-01-29
+
 ### Added
 
 - Parse manifest files with non-standard names